separate models into files
